move metafile to parseTorrent.ts
