--- roundtrip.mjs
@@ -1,4 +1,0 @@
-import parseTorrent from 'parse-torrent';
-import {readFileSync} from 'fs';
-const x = parseTorrent(readFileSync(0));
-parseTorrent(parseTorrent.toTorrentFile(x));
