+++ src/clients/QBittorrent.ts
@@ -3,11 +3,12 @@ import { FormData } from "formdata-polyfill/esm.min.js";
 import { unlink, writeFile } from "fs/promises";
 import fetch, { BodyInit, Response } from "node-fetch";
 import { tmpdir } from "os";
-import parseTorrent, { Metafile } from "parse-torrent";
+import { Metafile } from "../parseTorrent.js";
 import { join, posix } from "path";
 import { InjectionResult } from "../constants.js";
 import { CrossSeedError } from "../errors.js";
 import { Label, logger, logOnce } from "../logger.js";
+import { encodeTorrentFile } from "../parseTorrent.js";
 import { getRuntimeConfig } from "../runtimeConfig.js";
 import { Searchee } from "../searchee.js";
 import { isSingleFileTorrent } from "../torrent.js";
@@ -279,7 +280,7 @@ export default class QBittorrent implements TorrentClient {
 			if (await this.isInfoHashInClient(newTorrent.infoHash)) {
 				return InjectionResult.ALREADY_EXISTS;
 			}
-			const buf = parseTorrent.toTorrentFile(newTorrent);
+			const buf = encodeTorrentFile(newTorrent);
 			const filename = `${newTorrent.name}.cross-seed.torrent`;
 			const tempFilepath = join(tmpdir(), filename);
 			await writeFile(tempFilepath, buf, { mode: 0o644 });
