+++ src/clients/Transmission.ts
@@ -1,8 +1,9 @@
 import fetch, { Headers } from "node-fetch";
-import parseTorrent, { Metafile } from "parse-torrent";
+import { Metafile } from "../parseTorrent.js";
 import { InjectionResult } from "../constants.js";
 import { CrossSeedError } from "../errors.js";
 import { Label, logger } from "../logger.js";
+import { encodeTorrentFile } from "../parseTorrent.js";
 import { Result, resultOf, resultOfErr } from "../Result.js";
 import { getRuntimeConfig } from "../runtimeConfig.js";
 import { Searchee } from "../searchee.js";
@@ -174,9 +175,7 @@ export default class Transmission implements TorrentClient {
 				"torrent-add",
 				{
 					"download-dir": downloadDir,
-					metainfo: parseTorrent
-						.toTorrentFile(newTorrent)
-						.toString("base64"),
+					metainfo: encodeTorrentFile(newTorrent).toString("base64"),
 					paused: false,
 					labels: ["cross-seed"],
 				}
