+++ src/decide.ts
@@ -1,10 +1,11 @@
 import { existsSync, statSync, writeFileSync } from "fs";
-import parseTorrent, { Metafile } from "parse-torrent";
+import { Metafile } from "./parseTorrent.js";
 import path from "path";
 import { appDir } from "./configuration.js";
 import { MatchMode, Decision, TORRENT_CACHE_FOLDER } from "./constants.js";
 import { db } from "./db.js";
 import { Label, logger } from "./logger.js";
+import { encodeTorrentFile } from "./parseTorrent.js";
 import { Candidate } from "./pipeline.js";
 import { getRuntimeConfig } from "./runtimeConfig.js";
 import { File, getFiles, Searchee } from "./searchee.js";
@@ -159,7 +160,7 @@ function cacheTorrentFile(meta: Metafile): void {
 			TORRENT_CACHE_FOLDER,
 			`${meta.infoHash}.cached.torrent`
 		),
-		parseTorrent.toTorrentFile(meta)
+		encodeTorrentFile(meta)
 	);
 }
