+++ src/index.d.ts
@@ -3,50 +3,5 @@
 declare module "parse-torrent" {
 	export = ParseTorrent;
 	function ParseTorrent(torrentId: Buffer): ParseTorrent.Metafile;
-	namespace ParseTorrent {
-		function remote(
-			torrentId: string,
-			opts: unknown,
-			cb: (error: unknown, torrent: Metafile) => void
-		): void;
-
-		function toMagnetURI(obj: Metafile): string;
-
-		function toTorrentFile(parsed: Metafile): Buffer;
-		export interface FileListing {
-			length: number;
-			name: string;
-			offset: number;
-			path: string;
-		}
-
-		export interface Metafile {
-			announce: string[];
-			created: Date;
-			createdBy: string;
-			files: FileListing[];
-			info: {
-				files?: {
-					length: number;
-					path?: Buffer[];
-					"path.utf-8"?: Buffer[];
-				}[];
-				name?: Buffer;
-				"piece length": number;
-				pieces: Buffer;
-				private: number;
-			};
-			infoBuffer: Buffer;
-			infoHash: string;
-			infoHashBuffer: Buffer;
-			lastPieceLength: number;
-			length: number;
-			name: string;
-			pieceLength: number;
-			pieces: string[];
-			private: boolean;
-			urlList: string[];
-			comment: string;
-		}
-	}
+	namespace ParseTorrent {}
 }
